+++ utilities/table_properties_collectors/compact_on_deletion_collector.cc
@@ -4,10 +4,10 @@
 //  of patent rights can be found in the PATENTS file in the same directory.
 
 #ifndef ROCKSDB_LITE
-#include <memory>
-
-#include "rocksdb/utilities/table_properties_collectors.h"
 #include "utilities/table_properties_collectors/compact_on_deletion_collector.h"
+
+#include <memory>
+#include "rocksdb/utilities/table_properties_collectors.h"
 
 namespace rocksdb {
 
@@ -75,8 +75,9 @@ Status CompactOnDeletionCollector::AddUserKey(
   return Status::OK();
 }
 
-TablePropertiesCollector* CompactOnDeletionCollectorFactory::
-    CreateTablePropertiesCollector() {
+TablePropertiesCollector*
+CompactOnDeletionCollectorFactory::CreateTablePropertiesCollector(
+    TablePropertiesCollectorFactory::Context context) {
   return new CompactOnDeletionCollector(
       sliding_window_size_, deletion_trigger_);
 }